[wmco] fix validation for replicas
fixes validation for replicas field as it was being interpreted as
float while validating.
this was because `replicas` was defined as `int` in place of `int32`
which is causing the `operator-sdk generate crds` to not create a
`format` field in crd. This is a known issue in
operator-sdk 0.17.

revert back this change once we bump operator-sdk version to
0.18 or above

Bug 1835166

+++ test/e2e/create_test.go
@@ -68,7 +68,7 @@ func (tc *testContext) createWMC(replicas int, keyPair string) (*operator.Window
 		Spec: operator.WindowsMachineConfigSpec{
 			InstanceType: instanceType,
 			AWS:          &operator.AWS{CredentialAccountID: credentialAccountID, SSHKeyPair: keyPair},
-			Replicas:     replicas,
+			Replicas:     int32(replicas),
 		},
 	}
 	return wmco, framework.Global.Client.Create(context.TODO(), wmco,
